refactor(transfer): split monolithic transfer file
