+++ transfer/transfer.go
@@ -1,30 +1,17 @@
-// transfer/transfer.go
 package transfer
 
 import (
-	"bufio"
-	"bytes"
-	"context"
-	"crypto/sha256"
-	"encoding/binary"
 	"encoding/gob"
-	"encoding/hex"
-	"encoding/json"
 	"fmt"
 	"io"
-	"math"
 	"os"
 	"sync"
-	"syscall"
 	"time"
 
-	"github.com/zeebo/blake3"
 	"go.uber.org/zap"
 
 	"lvmsync_go/common"
 	"lvmsync_go/config"
-	"lvmsync_go/device"
-	"lvmsync_go/internal/blocksize"
 )
 
 // Transfer encapsulates transfer state shared across operations.
@@ -97,125 +84,6 @@ func SaveChecksumState(filename string, state *ChecksumState, logger *zap.Logger
 		return fmt.Errorf("encode checksum state: %w", err)
 	}
 	return nil
-}
-
-func prepareOutputWriter(out io.Writer, cfg *config.Config, logger *zap.Logger) (io.WriteCloser, *bufio.Writer, error) {
-	_ = logger
-	limitedOut := WrapRateLimitedWriter(out, cfg.SpeedLimit)
-	compWriter, err := NewCompressionWriter(limitedOut, cfg.Compress, cfg.CompressLevel, cfg.CompressConcurrency)
-	if err != nil {
-		return nil, nil, fmt.Errorf("failed to create compression writer: %w", err)
-	}
-	bufOut := bufio.NewWriter(compWriter)
-	return compWriter, bufOut, nil
-}
-
-func cleanupOutput(buf *bufio.Writer, w io.WriteCloser, logger *zap.Logger) {
-	if err := buf.Flush(); err != nil {
-		if logger != nil {
-			logger.Warn("Failed to flush output", zap.Error(err))
-		}
-	}
-	if err := w.Close(); err != nil {
-		if logger != nil {
-			logger.Warn("Failed to close writer", zap.Error(err))
-		}
-	}
-}
-
-func detectBlockSize(cfg *config.Config, source string, logger *zap.Logger) error {
-	if cfg.BlockSize != 0 {
-		return nil
-	}
-	bs, err := blocksize.Detect(source)
-	if err != nil {
-		return fmt.Errorf("auto-detect block size: %w", err)
-	}
-	cfg.BlockSize = bs
-	if logger != nil {
-		logger.Info("Auto-detected block size", zap.Int("block_size_bytes", cfg.BlockSize))
-	}
-	return nil
-}
-
-func gatherChangedRanges(snapshot string, blockSize int64, logger *zap.Logger) ([]Range, error) {
-	metadataDevice := GetMetadataDevice(snapshot)
-	if metadataDevice == "" {
-		return nil, fmt.Errorf("failed to determine metadata device from snapshot %s", snapshot)
-	}
-	ranges, err := GetDifferences(metadataDevice, blockSize)
-	if err != nil {
-		return nil, fmt.Errorf("error getting differences: %w", err)
-	}
-	if logger != nil {
-		logger.Info("Changed blocks determined", zap.Int("block_count", len(ranges)))
-	}
-	return ranges, nil
-}
-
-func prepareRanges(cfg *config.Config, snapshot, source string, logger *zap.Logger) ([]Range, error) {
-	if err := detectBlockSize(cfg, source, logger); err != nil {
-		return nil, err
-	}
-	if logger != nil {
-		logger.Info("Using block size", zap.Int("block_size_bytes", cfg.BlockSize))
-	}
-
-	_, blockSize, err := validateOffsetAndSize(0, cfg.BlockSize)
-	if err != nil {
-		return nil, err
-	}
-	ranges, err := gatherChangedRanges(snapshot, int64(blockSize), logger)
-	if err != nil {
-		return nil, err
-	}
-	return ranges, nil
-}
-
-func setupSourceFile(cfg *config.Config, source string) (*os.File, error) {
-	if cfg.ODirect {
-		tmp, err := os.Open(source)
-		if err != nil {
-			return nil, fmt.Errorf("failed to open source device %s: %w", source, err)
-		}
-		sector, err := DetectSectorSize(tmp)
-		_ = tmp.Close()
-		if err == nil && cfg.BlockSize%sector == 0 {
-			if f, direct, err := openFileODirect(source, os.O_RDONLY); err == nil && direct {
-				return f, nil
-			}
-		}
-	}
-	srcFile, err := os.Open(source)
-	if err != nil {
-		return nil, fmt.Errorf("failed to open source device %s: %w", source, err)
-	}
-	return srcFile, nil
-}
-
-func setupPipe(cfg *config.Config, logger *zap.Logger) ([2]int, func(), error) {
-	var pipeFds [2]int
-	cleanup := func() {}
-	if cfg.ZeroCopy {
-		if err := syscall.Pipe(pipeFds[:]); err != nil {
-			return pipeFds, cleanup, fmt.Errorf("failed to create pipe: %w", err)
-		}
-		cleanup = func() {
-			if closeErr := syscall.Close(pipeFds[0]); closeErr != nil {
-				if logger != nil {
-					logger.Warn("close pipe", zap.Int("fd", pipeFds[0]), zap.Error(closeErr))
-				}
-			}
-			if closeErr := syscall.Close(pipeFds[1]); closeErr != nil {
-				if logger != nil {
-					logger.Warn("close pipe", zap.Int("fd", pipeFds[1]), zap.Error(closeErr))
-				}
-			}
-		}
-	} else {
-		pipeFds[0], pipeFds[1] = -1, -1
-	}
-	return pipeFds, cleanup, nil
 }
 
 func (t *Transfer) dumpChangesCore(cfg *config.Config, snapshot, source string, out io.Writer, dedup DeduplicationStrategy, handshake string) (err error) {
@@ -312,577 +180,4 @@ func (t *Transfer) DumpChanges(cfg *config.Config, snapshot, source string, out 
 		t.Logger.Info("Deduplication disabled, performing full block transfer")
 	}
 	return t.DumpChangesSequential(cfg, snapshot, source, out)
-}
-
-var (
-	resumeMu     sync.Mutex
-	resumeBytes  int64
-	resumeLast   time.Time
-	resumeChunk  [32]byte
-	resumeOffset uint64
-	resumeLength uint32
-)
-
-type resumeState struct {
-	Transport         string `json:"transport"`
-	Compress          string `json:"compress"`
-	ChecksumAlgorithm string `json:"checksum_algorithm"`
-	DedupMode         string `json:"dedup_mode"`
-	Offset            uint64 `json:"offset"`
-	Length            uint32 `json:"length"`
-	Chunk             string `json:"chunk"`
-}
-
-type resumeCheckpoint struct {
-	Chunk  [32]byte
-	Offset uint64
-	Length uint32
-}
-
-func writeResumeState(cfg *config.Config, logger *zap.Logger, path string, offset uint64, length uint32, chunk [32]byte) {
-	rs := resumeState{
-		Transport:         cfg.Transport,
-		Compress:          cfg.Compress,
-		ChecksumAlgorithm: cfg.ChecksumAlgorithm,
-		DedupMode:         cfg.DedupMode,
-		Offset:            offset,
-		Length:            length,
-		Chunk:             hex.EncodeToString(chunk[:]),
-	}
-	data, err := json.Marshal(rs)
-	if err != nil {
-		if logger != nil {
-			logger.Warn("Failed to marshal resume state", zap.Error(err))
-		}
-		return
-	}
-	if err := os.WriteFile(path, data, 0o600); err != nil {
-		if logger != nil {
-			logger.Warn("Failed to update resume state", zap.Error(err))
-		}
-	}
-}
-
-func saveResumeState(cfg *config.Config, offset uint64, chunk [32]byte, size int64, logger *zap.Logger) {
-	if cfg.ResumeState == "" {
-		return
-	}
-	resumeMu.Lock()
-	defer resumeMu.Unlock()
-	if resumeLast.IsZero() {
-		resumeLast = time.Now()
-	}
-	resumeBytes += size
-	resumeChunk = chunk
-	resumeOffset = offset
-	resumeLength = uint32(size)
-	if (cfg.CheckpointBytes > 0 && resumeBytes >= int64(cfg.CheckpointBytes)) ||
-		(cfg.CheckpointInterval > 0 && time.Since(resumeLast) >= cfg.CheckpointInterval) {
-		writeResumeState(cfg, logger, cfg.ResumeState, resumeOffset, resumeLength, resumeChunk)
-		resumeBytes = 0
-		resumeLast = time.Now()
-	}
-}
-
-func finalizeResumeState(cfg *config.Config, logger *zap.Logger) {
-	if cfg.ResumeState == "" {
-		return
-	}
-	resumeMu.Lock()
-	defer resumeMu.Unlock()
-	if resumeLast.IsZero() {
-		return
-	}
-	if err := os.Remove(cfg.ResumeState); err != nil && !os.IsNotExist(err) {
-		if logger != nil {
-			logger.Warn("Failed to remove resume state", zap.Error(err))
-		}
-	}
-	resumeBytes = 0
-	resumeLast = time.Time{}
-}
-
-func readResumeState(cfg *config.Config, logger *zap.Logger) resumeCheckpoint {
-	var out resumeCheckpoint
-	if cfg.ResumeState == "" {
-		return out
-	}
-	data, err := os.ReadFile(cfg.ResumeState)
-	if err != nil {
-		return out
-	}
-	var rs resumeState
-	if err := json.Unmarshal(data, &rs); err != nil {
-		return out
-	}
-	if rs.Transport != cfg.Transport || rs.Compress != cfg.Compress ||
-		rs.ChecksumAlgorithm != cfg.ChecksumAlgorithm || rs.DedupMode != cfg.DedupMode {
-		return out
-	}
-	b, err := hex.DecodeString(rs.Chunk)
-	if err != nil || len(b) != 32 {
-		return out
-	}
-	copy(out.Chunk[:], b)
-	out.Offset = rs.Offset
-	out.Length = rs.Length
-	if logger != nil {
-		logger.Info("Resuming from chunk",
-			zap.String("resume_chunk", hex.EncodeToString(out.Chunk[:])),
-			zap.Uint64("resume_offset", out.Offset),
-			zap.Uint32("resume_length", out.Length))
-	}
-	return out
-}
-
-func findResumeIndex(cfg *config.Config, srcFile *os.File, ranges []Range, chk resumeCheckpoint, logger *zap.Logger) int {
-	if cfg.ResumeState == "" {
-		return 0
-	}
-	if cfg.DedupMode == "cdc" || cfg.DedupMode == "hybrid" {
-		next := chk.Offset + uint64(chk.Length)
-		for i := range ranges {
-			if next < ranges[i].Start {
-				return i
-			}
-			if next <= ranges[i].End {
-				ranges[i].Start = next
-				if logger != nil {
-					logger.Info("Resuming after offset", zap.Uint64("resume_offset", next))
-				}
-				return i
-			}
-		}
-		return len(ranges)
-	}
-	if chk.Chunk == [32]byte{} {
-		return 0
-	}
-	for i, r := range ranges {
-		offset, _, err := validateOffsetAndSize(r.Start, cfg.BlockSize)
-		if err != nil {
-			return 0
-		}
-		data, err := ReadBlockWithRetries(cfg, srcFile, offset, cfg.ZeroCopy, [2]int{-1, -1}, logger)
-		if err != nil {
-			continue
-		}
-		var sum [32]byte
-		if cfg.ChecksumAlgorithm == "sha256" {
-			sumArr := sha256.Sum256(data)
-			sum = [32]byte(sumArr)
-		} else {
-			sum = blake3.Sum256(data)
-		}
-		putBlockBuffer(data)
-		if sum == chk.Chunk {
-			if logger != nil {
-				logger.Info("Resuming after index", zap.Int("resume_index", i+1))
-			}
-			return i + 1
-		}
-	}
-	return 0
-}
-
-// DumpChangesParallel transfers changed blocks using multiple goroutines and updates resume state as blocks complete.
-
-func calculateTotalDataSize(ranges []Range) int64 {
-	var total uint64
-	for _, r := range ranges {
-		if r.End < r.Start {
-			continue
-		}
-		total += r.End - r.Start + 1
-	}
-	if total > uint64(math.MaxInt64) {
-		return math.MaxInt64
-	}
-	return int64(total)
-}
-
-func (t *Transfer) startParallelWorkers(cfg *config.Config, srcFile *os.File, ranges []Range, resumeStart int, logger *zap.Logger) <-chan *BlockResult {
-	numBlocks := len(ranges)
-	taskBuf := cfg.Parallel
-	if taskBuf < 1 {
-		taskBuf = 1
-	}
-	tasks := make(chan BlockTask, taskBuf)
-	results := make(chan *BlockResult, taskBuf)
-	for i := 0; i < cfg.Parallel; i++ {
-		t.workerWG.Add(1)
-		go worker(cfg, srcFile, tasks, results, t.workerWG, logger)
-	}
-
-	go func() {
-		for i := resumeStart; i < numBlocks; i++ {
-			tasks <- BlockTask{Index: i, R: ranges[i]}
-		}
-		close(tasks)
-	}()
-
-	go finalizeResults(t.workerWG, results)
-	return results
-}
-
-func (t *Transfer) DumpChangesParallel(cfg *config.Config, snapshot, source string, out io.Writer) (err error) {
-	if cfg.ZeroCopy {
-		if t.Logger != nil {
-			t.Logger.Warn("ZeroCopy mode enabled, falling back to sequential execution")
-		}
-		return t.DumpChangesSequential(cfg, snapshot, source, out)
-	}
-
-	ranges, err := prepareRanges(cfg, snapshot, source, t.Logger)
-	if err != nil {
-		return err
-	}
-
-	totalDataSize := calculateTotalDataSize(ranges)
-	if err = writeParallelHandshake(cfg, out); err != nil {
-		return err
-	}
-
-	compWriter, bufOut, err := prepareOutputWriter(out, cfg, t.Logger)
-	if err != nil {
-		return err
-	}
-	defer cleanupOutput(bufOut, compWriter, t.Logger)
-
-	srcFile, err := setupSourceFile(cfg, source)
-	if err != nil {
-		return err
-	}
-	defer common.CloseWithErr(srcFile, &err, "close source file")
-
-	checkpoint := readResumeState(cfg, t.Logger)
-	resumeStart := findResumeIndex(cfg, srcFile, ranges, checkpoint, t.Logger)
-	results := t.startParallelWorkers(cfg, srcFile, ranges, resumeStart, t.Logger)
-
-	startTime := time.Now()
-	checksum := GetChecksumStrategy(cfg.ChecksumAlgorithm)
-	var totalBytesTransferred int64
-	var finalDigest []byte
-	totalBytesTransferred, finalDigest, err = processParallelResults(cfg, results, bufOut, checksum, totalDataSize, startTime, t.Logger)
-	if err != nil {
-		return err
-	}
-	finalizeProgress(cfg, t.Logger)
-	logParallelSummary(t.Logger, totalBytesTransferred, startTime)
-	finalizeResumeState(cfg, t.Logger)
-	if len(finalDigest) > 0 && t.Logger != nil {
-		t.Logger.Info("final checksum", zap.String("final_digest", fmt.Sprintf("%x", finalDigest)))
-	}
-	if t.Logger != nil {
-		_ = t.Logger.Sync()
-	}
-	return nil
-}
-
-func readBlockHeader(reader *bufio.Reader, headerBuf []byte, verify bool, checksum ChecksumStrategy) (uint64, uint32, []byte, error) {
-	_, err := io.ReadFull(reader, headerBuf)
-	if err == io.EOF || err == io.ErrUnexpectedEOF {
-		return 0, 0, nil, io.EOF
-	}
-	if err != nil {
-		return 0, 0, nil, fmt.Errorf("failed to read chunk header: %w", err)
-	}
-
-	offset := binary.BigEndian.Uint64(headerBuf[0:8])
-	if offset > math.MaxInt64 {
-		return 0, 0, nil, fmt.Errorf("offset %d overflows int64", offset)
-	}
-	chunkSize := binary.BigEndian.Uint32(headerBuf[8:12])
-
-	var transmittedSum []byte
-	if verify {
-		transmittedSum = make([]byte, checksum.Size())
-		copy(transmittedSum, headerBuf[12:])
-	}
-
-	return offset, chunkSize, transmittedSum, nil
-}
-
-func readBlockData(cfg *config.Config, reader io.Reader, chunkSize uint32) ([]byte, error) {
-	if chunkSize == 0 {
-		return nil, nil
-	}
-	size := int(chunkSize)
-	var data []byte
-	if cfg.ODirect {
-		data = getAlignedBlockBuffer(size)
-	} else {
-		data = getBlockBuffer(size)
-	}
-	if _, err := io.ReadFull(reader, data); err != nil {
-		if cfg.ODirect {
-			putAlignedBlockBuffer(data)
-		} else {
-			putBlockBuffer(data)
-		}
-		if err == io.EOF || err == io.ErrUnexpectedEOF {
-			return nil, io.EOF
-		}
-		return nil, fmt.Errorf("failed to read chunk data: %w", err)
-	}
-	return data, nil
-}
-
-func verifyChecksum(verify bool, checksum ChecksumStrategy, data, transmitted []byte, offset uint64) error {
-	if !verify {
-		return nil
-	}
-	computed := checksum.Compute(data)
-	if !bytes.Equal(transmitted, computed) {
-		return fmt.Errorf("checksum mismatch at offset %d", offset)
-	}
-	return nil
-}
-
-func writeData(destFile *os.File, offset uint64, data []byte, logger *zap.Logger) error {
-	if offset > math.MaxInt64 {
-		return fmt.Errorf("offset %d overflows int64", offset)
-	}
-	if _, err := destFile.Seek(int64(offset), io.SeekStart); err != nil {
-		if logger != nil {
-			logger.Warn("Seek error", zap.Uint64("offset", offset), zap.Error(err))
-		}
-		return nil
-	}
-	if _, err := destFile.Write(data); err != nil {
-		return fmt.Errorf("failed to write data at offset %d: %w", offset, err)
-	}
-	return nil
-}
-
-func processBlock(
-	cfg *config.Config,
-	destFile *os.File,
-	dedup DeduplicationStrategy,
-	verify bool,
-	checksum ChecksumStrategy,
-	offset uint64,
-	transmitted, data []byte,
-	chunkSize uint32,
-	logger *zap.Logger,
-) (bool, error) {
-	if offset > math.MaxInt64 {
-		return false, fmt.Errorf("offset %d overflows int64", offset)
-	}
-	if err := verifyChecksum(verify, checksum, data, transmitted, offset); err != nil {
-		return false, err
-	}
-	if chunkSize == 0 || isAllZero(data) {
-		if err := punchHole(destFile, offset, cfg.BlockSize); err != nil {
-			zero := getAlignedBlockBuffer(cfg.BlockSize)
-			if err := writeData(destFile, offset, zero, logger); err != nil {
-				putAlignedBlockBuffer(zero)
-				return false, err
-			}
-			putAlignedBlockBuffer(zero)
-		}
-		return true, nil
-	}
-	if dedup != nil {
-		intOffset := int64(offset)
-		if !dedup.ShouldTransfer(intOffset, data) {
-			return false, nil
-		}
-		dedup.RecordTransfer(intOffset, data)
-	}
-	if err := writeData(destFile, offset, data, logger); err != nil {
-		return false, err
-	}
-	return true, nil
-}
-
-func applyBlocks(cfg *config.Config, reader *bufio.Reader, destFile *os.File, dedup DeduplicationStrategy, verify bool, checksum ChecksumStrategy, logger *zap.Logger) (int64, error) {
-	var totalBytes int64
-	var sinceSync int64
-	headerLen := 12
-	if verify {
-		headerLen += checksum.Size()
-	}
-	headerBuf := make([]byte, headerLen)
-	for {
-		offset, chunkSize, transmittedSum, err := readBlockHeader(reader, headerBuf, verify, checksum)
-		if err == io.EOF {
-			break
-		}
-		if err != nil {
-			return totalBytes, err
-		}
-
-		data, err := readBlockData(cfg, reader, chunkSize)
-		if err == io.EOF {
-			break
-		}
-		if err != nil {
-			return totalBytes, err
-		}
-		written, err := processBlock(cfg, destFile, dedup, verify, checksum, offset, transmittedSum, data, chunkSize, logger)
-		if cfg.ODirect {
-			if data != nil {
-				putAlignedBlockBuffer(data)
-			}
-		} else if data != nil {
-			putBlockBuffer(data)
-		}
-		if err != nil {
-			return totalBytes, err
-		}
-		if written {
-			totalBytes += int64(chunkSize)
-			sinceSync += int64(chunkSize)
-			if cfg.SyncIntervalBytes > 0 && sinceSync >= int64(cfg.SyncIntervalBytes) {
-				if err := fdatasyncFile(destFile); err != nil {
-					return totalBytes, err
-				}
-				sinceSync = 0
-			}
-		}
-	}
-	return totalBytes, nil
-}
-
-func (t *Transfer) processDumpDataCore(cfg *config.Config, in io.Reader, destPath string, dedup DeduplicationStrategy, verify bool) (err error) {
-	bufReader := bufio.NewReader(in)
-	var hs common.Handshake
-	hs, err = readAndValidateHandshake(cfg, bufReader, dedup, verify)
-	if err != nil {
-		return err
-	}
-
-	decReader, err := NewDecompressionReader(bufReader, hs.Compress, cfg.CompressConcurrency)
-	if err != nil {
-		return fmt.Errorf("failed to create decompression reader: %w", err)
-	}
-	defer func() {
-		if closeErr := decReader.Close(); closeErr != nil && t.Logger != nil {
-			t.Logger.Warn("Failed to close decompression reader", zap.Error(closeErr))
-		}
-	}()
-
-	reader := bufio.NewReader(decReader)
-
-	if cfg.DeviceUUID != "" {
-		uuid, err2 := device.GetUUID(context.Background(), destPath)
-		if err2 != nil {
-			return fmt.Errorf("read destination uuid: %w", err2)
-		}
-		if uuid != cfg.DeviceUUID {
-			return fmt.Errorf("destination device uuid %s does not match expected %s", uuid, cfg.DeviceUUID)
-		}
-	}
-	mounted, err2 := device.IsMountedRW(destPath)
-	if err2 != nil {
-		return fmt.Errorf("check mount status: %w", err2)
-	}
-	if mounted && !cfg.Force {
-		return fmt.Errorf("destination device %s is mounted read-write", destPath)
-	}
-
-	var destFile *os.File
-	if cfg.ODirect {
-		tmp, err2 := os.Open(destPath)
-		if err2 != nil {
-			return fmt.Errorf("failed to open destination device %s: %w", destPath, err2)
-		}
-		sector, err2 := DetectSectorSize(tmp)
-		_ = tmp.Close()
-		if err2 == nil && cfg.BlockSize%sector == 0 {
-			if f, direct, err2 := openFileODirect(destPath, os.O_RDWR); err2 == nil && direct {
-				destFile = f
-			}
-		}
-	}
-	if destFile == nil {
-		destFile, err = os.OpenFile(destPath, os.O_RDWR, 0)
-		if err != nil {
-			return fmt.Errorf("failed to open destination device %s: %w", destPath, err)
-		}
-	}
-	defer common.CloseWithErr(destFile, &err, "close destination device")
-
-	startTime := time.Now()
-	checksum := GetChecksumStrategy(cfg.ChecksumAlgorithm)
-	var totalBytes int64
-	totalBytes, err = applyBlocks(cfg, reader, destFile, dedup, verify, checksum, t.Logger)
-	if err != nil {
-		return err
-	}
-
-	elapsed := time.Since(startTime).Seconds()
-	if t.Logger != nil {
-		t.Logger.Info("Applied changes",
-			zap.Int64("bytes", totalBytes),
-			zap.Float64("seconds", elapsed),
-			zap.Float64("MB/s", float64(totalBytes)/elapsed/1048576.0))
-	}
-	return nil
-}
-
-// ProcessDumpDataWithDeduplication applies a dump stream to destPath using the given dedup strategy without checksum verification, updating the strategy's state.
-func (t *Transfer) ProcessDumpDataWithDeduplication(cfg *config.Config, in io.Reader, destPath string, dedup DeduplicationStrategy) error {
-	return t.processDumpDataCore(cfg, in, destPath, dedup, false)
-}
-
-// ProcessDumpData applies a dump stream to destPath with checksum verification for each block before writing.
-func (t *Transfer) ProcessDumpData(cfg *config.Config, in io.Reader, destPath string) error {
-	return t.processDumpDataCore(cfg, in, destPath, nil, true)
-}
-
-func openApplyReader(applyFile string) (io.ReadCloser, error) {
-	if applyFile == "-" {
-		return io.NopCloser(os.Stdin), nil
-	}
-	f, err := os.Open(applyFile)
-	if err != nil {
-		return nil, fmt.Errorf("failed to open apply file %s: %w", applyFile, err)
-	}
-	return f, nil
-}
-
-func (t *Transfer) applyData(cfg *config.Config, in io.Reader, destDevice string) error {
-	if cfg.DeviceUUID != "" {
-		uuid, err := device.GetUUID(context.Background(), destDevice)
-		if err != nil {
-			return fmt.Errorf("read destination uuid: %w", err)
-		}
-		if uuid != cfg.DeviceUUID {
-			return fmt.Errorf("destination device uuid %s does not match expected %s", uuid, cfg.DeviceUUID)
-		}
-	}
-	mounted, err := device.IsMountedRW(destDevice)
-	if err != nil {
-		return fmt.Errorf("check mount status: %w", err)
-	}
-	if mounted && !cfg.Force {
-		return fmt.Errorf("destination device %s is mounted read-write", destDevice)
-	}
-	dedup := NewDeduplicationStrategy(cfg, t.Logger)
-	if dedup != nil {
-		if t.Logger != nil {
-			t.Logger.Info("Applying deduplication during restore", zap.String("strategy", cfg.DedupStrategy))
-		}
-		defer func() {
-			if err := dedup.SaveState(); err != nil {
-				if t.Logger != nil {
-					t.Logger.Error("Failed to save dedup state", zap.Error(err))
-				}
-			}
-		}()
-		return t.ProcessDumpDataWithDeduplication(cfg, in, destDevice, dedup)
-	}
-	return t.ProcessDumpData(cfg, in, destDevice)
-}
-
-// RunApply reads a dump file or stdin and writes the data to destDevice.
-func (t *Transfer) RunApply(cfg *config.Config, applyFile, destDevice string) (err error) {
-	rc, err := openApplyReader(applyFile)
-	if err != nil {
-		return err
-	}
-	defer common.CloseWithErr(rc, &err, "close apply file")
-	return t.applyData(cfg, rc, destDevice)
 }
